% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-edges.R
\name{geom_edgetext}
\alias{geom_edgetext}
\title{Label the edges of a network.}
\usage{
geom_edgetext(mapping = NULL, data = NULL, position = "identity",
  parse = FALSE, show.legend = NA, inherit.aes = TRUE, ..., nudge_x = 0,
  nudge_y = 0, check_overlap = FALSE)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{geom_text}}}

\item{data}{see \code{\link[ggplot2]{geom_text}}}

\item{position}{see \code{\link[ggplot2]{geom_text}}}

\item{parse}{see \code{\link[ggplot2]{geom_text}}}

\item{show.legend}{see \code{\link[ggplot2]{geom_text}}}

\item{inherit.aes}{see \code{\link[ggplot2]{geom_text}}}

\item{...}{see \code{\link[ggplot2]{geom_text}}}

\item{nudge_x}{see \code{\link[ggplot2]{geom_text}}}

\item{nudge_y}{see \code{\link[ggplot2]{geom_text}}}

\item{check_overlap}{see \code{\link[ggplot2]{geom_text}}}
}
\description{
All arguments to this geom are identical to those of
\code{\link[ggplot2]{geom_text}}. The \code{fill} aesthetic, which defaults
to \code{"white"}, can be used to set the background color of the labels. The
labels will be drawn at mid-edge.
}

