% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-nodes.R
\docType{data}
\name{geom_nodes}
\alias{StatNodes}
\alias{geom_nodes}
\title{Draw the nodes of a network.}
\usage{
geom_nodes(mapping = NULL, data = NULL, stat = "nodes",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{geom_point}}}

\item{data}{see \code{\link[ggplot2]{geom_point}}}

\item{stat}{see \code{\link[ggplot2]{geom_point}}}

\item{position}{see \code{\link[ggplot2]{geom_point}}}

\item{na.rm}{see \code{\link[ggplot2]{geom_point}}}

\item{show.legend}{see \code{\link[ggplot2]{geom_point}}}

\item{inherit.aes}{see \code{\link[ggplot2]{geom_point}}}

\item{...}{see \code{\link[ggplot2]{geom_point}}}
}
\description{
All arguments to this function are identical to those of
\code{\link[ggplot2]{geom_point}}.
}
\keyword{datasets}

