% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggnetwork.R
\name{ggnetwork}
\alias{ggnetwork}
\title{Convert a network object to a data frame}
\usage{
ggnetwork(x, layout = "fruchtermanreingold",
  arrow.gap = ifelse(network::is.directed(x), 0.025, 0), ...)
}
\arguments{
\item{x}{an object of class \code{\link[network]{network}}, or any object
that can be coerced to this class, such as an adjacency or incidence matrix,
or an edge list: see \link[network]{edgeset.constructors} and
\link[network]{network} for details. If the object is of class
\code{\link[igraph:igraph-package]{igraph}} and the
\code{\link[intergraph:intergraph-package]{intergraph}} package is installed,
it will be used to convert the object: see
\code{\link[intergraph]{asNetwork}} for details.}

\item{layout}{a network layout supplied by \code{\link[sna]{gplot.layout}}}

\item{arrow.gap}{a parameter that will shorten the network edges in order to
avoid overplotting edge arrows and nodes; defaults to \code{0} when the
network is undirected (no edge shortening), or to \code{0.025} when the
network is directed.}

\item{...}{other parameters that will get passed to the network layout
algorithm as a list}
}
\description{
See the vignette for a description of both this function and the rest of the
\code{ggnetwork} package.
}
\examples{
if (require(network)) {
  data(emon)
  ggnetwork(emon[[1]])
  ggnetwork(emon[[1]], layout = "target", niter = 100)
}
}

