% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scale_safely}
\alias{scale_safely}
\title{Rescale `x` to (0, 1), except if `x` is constant}
\usage{
scale_safely(x, scale = diff(range(x)))
}
\arguments{
\item{x}{a vector to rescale}

\item{scale}{the scale on which to rescale the vector}
}
\value{
The rescaled vector, coerced to a vector if necessary.
}
\description{
Discussed in PR #32.
}
\author{
Kipp Johnson
}
