% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_blank}
\alias{theme_blank}
\title{Blank ggplot2 theme, suited for plotting networks.}
\usage{
theme_blank(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{other \code{\link[ggplot2]{theme}} arguments}
}
\description{
A \code{ggplot2} theme without lines, borders, axis text or titles, suited
for plotting networks.
}
