% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_facet}
\alias{theme_facet}
\title{Blank ggplot2 theme with a panel border.}
\usage{
theme_facet(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{other \code{\link{theme}} arguments}
}
\description{
A variation of \code{\link{theme_blank}} that adds a panel border to the
plot, which is often suitable for plotting faceted networks.
}

