% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-igraph.R
\name{fortify.igraph}
\alias{fortify.igraph}
\title{Fortify method for networks of class \code{\link[igraph:igraph-package]{igraph}}}
\usage{
\method{fortify}{igraph}(model, ...)
}
\arguments{
\item{model}{an object of class \code{\link[igraph:igraph-package]{igraph}},
which will be converted to an object of class \code{\link[network]{network}}
with the \code{\link[intergraph]{asNetwork}} function before being passed to
the \code{\link{fortify.network}} function.}

\item{...}{additional parameters for the \code{fortify.network} function; see
\code{\link{fortify.network}}.}
}
\description{
This function requires the \code{\link[intergraph:intergraph-package]{intergraph}}
package to be installed.
}

