% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-edges.R
\docType{data}
\name{geom_edges}
\alias{StatEdges}
\alias{geom_edges}
\title{Draw the edges of a network.}
\usage{
geom_edges(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", arrow = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{geom_segment}}}

\item{data}{see \code{\link[ggplot2]{geom_segment}}}

\item{stat}{see \code{\link[ggplot2]{geom_segment}}}

\item{position}{see \code{\link[ggplot2]{geom_segment}}}

\item{arrow}{see \code{\link[ggplot2]{geom_segment}}}

\item{na.rm}{see \code{\link[ggplot2]{geom_segment}}}

\item{show.legend}{see \code{\link[ggplot2]{geom_segment}}}

\item{inherit.aes}{see \code{\link[ggplot2]{geom_segment}}}

\item{...}{see \code{\link[ggplot2]{geom_segment}}}
}
\description{
All arguments to this function are identical to those of
\code{\link[ggplot2]{geom_segment}}, including \code{arrow}.
}
\keyword{datasets}

