% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-edges.R
\name{geom_edges}
\alias{geom_edges}
\title{Draw the edges of a network.}
\usage{
geom_edges(mapping = NULL, data = NULL, position = "identity",
  arrow = NULL, na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  ...)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{geom_segment}}}

\item{data}{see \code{\link[ggplot2]{geom_segment}}}

\item{position}{see \code{\link[ggplot2]{geom_segment}}}

\item{arrow}{see \code{\link[ggplot2]{geom_segment}}}

\item{na.rm}{see \code{\link[ggplot2]{geom_segment}}}

\item{show.legend}{see \code{\link[ggplot2]{geom_segment}}}

\item{inherit.aes}{see \code{\link[ggplot2]{geom_segment}}}

\item{...}{see \code{\link[ggplot2]{geom_segment}}}
}
\description{
All arguments to this geom are identical to those of
\code{\link[ggplot2]{geom_segment}}, including \code{arrow}, which is useful
to plot directed networks in conjunction with the \code{arrow.gap} argument
of \code{\link{ggnetwork}}.
}

