% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-nodes.R
\name{geom_nodetext}
\alias{geom_nodetext}
\title{Label the nodes of a network.}
\usage{
geom_nodetext(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, parse = FALSE, show.legend = NA, inherit.aes = TRUE,
  ..., nudge_x = 0, nudge_y = 0, check_overlap = FALSE)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{geom_text}}}

\item{data}{see \code{\link[ggplot2]{geom_text}}}

\item{position}{see \code{\link[ggplot2]{geom_text}}}

\item{na.rm}{see \code{\link[ggplot2]{geom_text}}}

\item{parse}{see \code{\link[ggplot2]{geom_text}}}

\item{show.legend}{see \code{\link[ggplot2]{geom_text}}}

\item{inherit.aes}{see \code{\link[ggplot2]{geom_text}}}

\item{...}{see \code{\link[ggplot2]{geom_text}}}

\item{nudge_x}{see \code{\link[ggplot2]{geom_text}}}

\item{nudge_y}{see \code{\link[ggplot2]{geom_text}}}

\item{check_overlap}{see \code{\link[ggplot2]{geom_text}}}
}
\description{
All arguments to this geom are identical to those of
\code{\link[ggplot2]{geom_text}}.
}

