% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork.R
\name{ggnetwork}
\alias{ggnetwork}
\title{ggnetwork}
\usage{
ggnetwork(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[network]{network}}, or any object
that can be coerced to this class, such as an adjacency or incidence matrix,
or an edge list: see \link[network]{edgeset.constructors} and
\link[network]{network} for details. If the object is of class
\code{\link[igraph:igraph-package]{igraph}} and the
\code{\link[intergraph:intergraph-package]{intergraph}} package is installed,
it will be used to convert the object: see
\code{\link{fortify.igraph}} for details.}

\item{...}{arguments passed to the \code{\link{fortify.network}} function.}
}
\description{
A wrapper for the \code{\link{fortify.network}} and
\code{\link{fortify.igraph}} functions.
}

